PR後の内容修正
