+++ app/Http/Controllers/FavoriteCommentsController.php
@@ -4,20 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 
-class FavoriteController extends Controller
+class FavoriteCommentsController extends Controller
 {
-    public function storePost($postId)
-    {
-        \Auth::user()->favoritePost($postId);
-        return back();
-    }
-
-    public function destroyPost($postId)
-    {
-        \Auth::user()->unfavoritePost($postId);
-        return back();
-    }
-
     public function storeComment($commentId)
     {
         \Auth::user()->favoriteComment($commentId);
