+++ app/Post.php
@@ -21,6 +21,6 @@ class Post extends Model
 
     public function favoritePostUsers()
     {
-        return $this->belongsToMany(User::class, 'favorites', 'favorite_post_id', 'favorite_user_id')->withTimestamps();
+        return $this->belongsToMany(User::class, 'favorite_posts', 'post_id', 'user_id')->withTimestamps();
     }
 }
